requirements.txt
django
gunicorn
python-dotenv
psycopg2-binary
psycopg2
whitenoise
pytz
djangorestframework